task: adding logout option
