+++ handle_login.go
@@ -6,15 +6,6 @@ import (
 
 // handleLogin enpoint renders the handleLogin page to enter the user
 // identifier.
-func (m *maildoor) handleLogin(w http.ResponseWriter, r *http.Request) {
-	data := atempt{
-		Logo: m.logoURL,
-	}
-
-	err := m.render(w, data, "layout.html", "handle_login.html")
-	if err != nil {
-		m.httpError(w, err)
-
-		return
-	}
+func (m *maildoor) handleLogout(w http.ResponseWriter, r *http.Request) {
+	m.logout(w, r)
 }
