+++ handle_logout.go
@@ -1,0 +1,20 @@
+package maildoor
+
+import (
+	"net/http"
+)
+
+// handleLogin enpoint renders the handleLogin page to enter the user
+// identifier.
+func (m *maildoor) handleLogin(w http.ResponseWriter, r *http.Request) {
+	data := atempt{
+		Logo: m.logoURL,
+	}
+
+	err := m.render(w, data, "layout.html", "handle_login.html")
+	if err != nil {
+		m.httpError(w, err)
+
+		return
+	}
+}
